created CRUD view for insurance policies

+++ InsurancePoliciesManager/Gap.IPM.Application/InsurancePolicies/Commands/UpdateInsurancePolicy/UpdateInsurancePolicyCommand.cs
@@ -15,7 +15,7 @@ namespace Gap.IPM.Application.InsurancePolicies.Commands.UpdateInsurancePolicy
         public string Name { get; set; }
         public string Description { get; set; }
         public int CoverageTypeId { get; set; }
-        public int Coverage { get; set; }
+        public decimal Coverage { get; set; }
         public DateTime CoverageStart { get; set; }
         public int CoveragePeriod { get; set; }
         public Int64 PolicyValue { get; set; }
